pyproject.toml
[project]
name = "tools"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = ["structlog>=25.5.0"]

[dependency-groups]
dev = ["pre-commit>=4.5.1", "ruff>=0.14.10", "ty>=0.0.4", "zizmor>=1.19.0"]
tool-convert-page-images = [
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "types-pyyaml>=6.0.12.20250915",
]

ci-detect-tools-changes = ["pyyaml>=6.0.3"]
tool-dev = [
    "playwright>=1.57.0",
]
